+++ derive/src/derive/std/error.rs
@@ -9,6 +9,11 @@ pub(crate) fn derive(data: &Data) -> Result<TokenStream> {
     let root = std_root();
     let trait_ = quote!(#root::error::Error);
 
+    #[cfg(all(stable_1_33, not(feature = "error_cause")))]
+    let cause = TokenStream::new();
+    #[cfg(any(not(stable_1_33), feature = "error_cause"))]
+    let cause = quote!(#[allow(deprecated)] fn cause(&self) -> #root::option::Option<&dyn (#trait_)>;);
+
     data.impl_trait_with_capacity(
         2,
         syn::parse2(trait_.clone())?,
@@ -16,8 +21,7 @@ pub(crate) fn derive(data: &Data) -> Result<TokenStream> {
         syn::parse2(quote! {
             trait Error {
                 fn description(&self) -> &str;
-                #[allow(deprecated)]
-                fn cause(&self) -> #root::option::Option<&dyn (#trait_)>;
+                #cause
                 fn source(&self) -> #root::option::Option<&(dyn (#trait_) + 'static)>;
             }
         })?,