+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+* Change `#[enum_derive(Error)]` implementation
+  In Rust 1.33, `Error::cause` is deprecated. In the new implementation, `Error::cause` is optional for Rust 1.33 and later. In versions less than 1.33, `Error::cause` is always implemented.
+
 # 0.1.2 - 2018-12-15
 
 * Move features of derive/utils to [derive_utils](https://github.com/taiki-e/derive_utils)